Update Bindings to Use Cache

+++ src/bindings.h
@@ -29,7 +29,7 @@ extern "C"
     uint64_t get_token(const char *uri, const char **options);
     void surrender_token(uint64_t token);
     int read_data(uint64_t token,
-                  double src_window[4],
+                  int src_window[4],
                   int dst_window[2],
                   int band_number,
                   GDALDataType type,
